+++ examples/ga/onemax_multidemic.py
@@ -40,7 +40,7 @@ def evalOneMax(individual):
     return sum(individual),
 
 toolbox.register("evaluate", evalOneMax)
-toolbox.register("mate", tools.cxTwoPoints)
+toolbox.register("mate", tools.cxTwoPoint)
 toolbox.register("mutate", tools.mutFlipBit, indpb=0.05)
 toolbox.register("select", tools.selTournament, tournsize=3)
 toolbox.register("migrate", tools.migRing, k=5, selection=tools.selBest,